+++ internal/domain/repository/photo.go
@@ -11,9 +11,9 @@ type PhotoRepository struct {
 }
 
 type IPhotoRepository interface {
-	GetByID(id uint64) (entity.PhotoView, error)
-	FindByUserID(userID uint64) ([]*entity.Photo, error)
-	FindLatestByUserID(userID uint64) (entity.PhotoViewModel, error)
+	GetByID(id int) (entity.PhotoView, error)
+	FindByUserID(userID int) ([]*entity.Photo, error)
+	FindLatestByUserID(userID int) (entity.PhotoViewModel, error)
 	Save(photo *entity.Photo) (*entity.Photo, error)
 }
 
@@ -25,19 +25,19 @@ func NewPhoto() *entity.Photo {
 	return &entity.Photo{}
 }
 
-func (r *PhotoRepository) GetByID(id uint64) (entity.PhotoView, error) {
+func (r *PhotoRepository) GetByID(id int) (entity.PhotoView, error) {
 	var photo entity.PhotoView
 	err := r.db.First(&photo, id).Error
 	return photo, err
 }
 
-func (r *PhotoRepository) FindByUserID(userID uint64) ([]*entity.Photo, error) {
+func (r *PhotoRepository) FindByUserID(userID int) ([]*entity.Photo, error) {
 	var photos []*entity.Photo
 	err := r.db.Where("user_id = ?", userID).Find(&photos).Error
 	return photos, err
 }
 
-func (r *PhotoRepository) FindLatestByUserID(userID uint64) (entity.PhotoViewModel, error) {
+func (r *PhotoRepository) FindLatestByUserID(userID int) (entity.PhotoViewModel, error) {
 	var photo entity.PhotoViewModel
 	err := r.db.Where("user_id = ?", userID).Order("created_at desc").First(&photo).Error
 	return photo, err
